major update

+++ worldspace2/worldspace2.cpp
@@ -187,7 +187,7 @@ int main(const int argc, char** argv)
 		Controls controls;
 		INI ini{ Global.myPath };
 
-		point gridSize{ ini.getvs_cast<position>("game", "iGridSizeX", str::stoll).value_or(Global.DEFAULT_SIZE_X), ini.getvs_cast<position>("game", "iGridSizeY", str::stoll).value_or(Global.DEFAULT_SIZE_Y) };
+		point gridSize{ ini.getvs_cast<position>("game", "iGridSizeX", str::stoi).value_or(Global.DEFAULT_SIZE_X), ini.getvs_cast<position>("game", "iGridSizeY", str::stoi).value_or(Global.DEFAULT_SIZE_Y) };
 
 		std::mutex mutex;
